added the tested versions

+++ ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: jheiskan <jheiskan@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/04 10:23:37 by jheiskan          #+#    #+#             */
-/*   Updated: 2021/11/15 22:44:22 by jheiskan         ###   ########.fr       */
+/*   Updated: 2021/11/12 16:37:19 by jheiskan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /* NO error checking version, as in the original */
 int	ft_atoi(const char *str)
 {
-	int				i;
-	int				sign;
-	long long int	res;
+	int	i;
+	int	sign;
+	int	res;
 
 	i = 0;
 	sign = 1;
